Fix ie11 fill (#1018)
* Fix for 'update' of null error in Sentry

We were trying to update Perfect Scrollbar after 400ms timeout. If the was user already left the page it caused an error. Now we only update it if the ref to perfect scrollbar is sill present on page after timeout.

See:

https://sentry.io/organizations/iterative/issues/1529236771/
https://sentry.io/organizations/iterative/issues/1525523962/

And others like that.

* Replace Array.fill with lodash to support IE11

Another Sentry error https://sentry.io/organizations/iterative/issues/1521458935/