Add service account information for GCloud
This commit adds some additional instructions to successfully set up a service account and configure a DVC command line remote that is pointing to a GCS bucket to use said service account. 

Its unclear to me if the remote setup needs the separate `GOOGLE_APPLICATION_CREDENTIALS` env call - but it worked here.