Move cache pruning into a gatsby hook and fix prefixes.
I believe the crux of what keeps this PR from working is that `public` and
`.cache` the folders were being uploaded to `public` and `,cache` the S3
prefixes when the site is deployed from the root prefix. This change alters how
the cache directories are stored: now as tuple pairs with the local directory in
slot 0 and a string that's appended to the base prefix in slot 2. `public` work
with the root prefix and the `.cache` folder works with the root prefix with
`-cache` appended to it such that it's another "folder" that doesn't get
downloaded with the root prefix.

I find the tuple syntax more concise, but I'm open to changing it if we prefer
objects for better grokking at a glance.

This change also moves the responsibility of cache culling into an `onPostBuild`
Gatsby hook. This means we don't have to parse the sitemap XML after the fact,
and as such the XML parser used to do it is removed. Since this replaces
`cleanUpPageData`, it also invalidates the whole `page-data-utils` file which
has been removed.