Rebuilds caching (#1182)
* Store .cache and public fully on S3

* Do not cache public and .cache folders on Heroku

* Clean up outdated page-data on rebuilds

* Get rootDir with process.cwd instead of __dirname

This way the code doesn't depend on the source file's relative directory.

* Parallelize uploadAllToS3, downloadAllFromS3, and clean.

I'm not sure if there's any rate limits in the stack that will cause this to break, but on first blush this looks like a quick and easy improvement.

The cached folders are also now accessed by a shared array that is hard-coded for now. This both makes them easier to work with uniformly and open to sourcing directories from elsewhere in the future.

* Fix a broken path.join

* Remove unreachable return

* Improve grammar in log messages

* Further grammar changes in logging

* Move cache pruning into a gatsby hook and fix prefixes.

I believe the crux of what keeps this PR from working is that `public` and
`.cache` the folders were being uploaded to `public` and `,cache` the S3
prefixes when the site is deployed from the root prefix. This change alters how
the cache directories are stored: now as tuple pairs with the local directory in
slot 0 and a string that's appended to the base prefix in slot 2. `public` work
with the root prefix and the `.cache` folder works with the root prefix with
`-cache` appended to it such that it's another "folder" that doesn't get
downloaded with the root prefix.

I find the tuple syntax more concise, but I'm open to changing it if we prefer
objects for better grokking at a glance.

This change also moves the responsibility of cache culling into an `onPostBuild`
Gatsby hook. This means we don't have to parse the sitemap XML after the fact,
and as such the XML parser used to do it is removed. Since this replaces
`cleanUpPageData`, it also invalidates the whole `page-data-utils` file which
has been removed.

* Remove left over console.log statement

* Refactor S3 calls from deploy-with-s3 to s3-utils

* Re-add functionality for moving index.html to project root.

* Add a clarifying comment on the 404 HTML file moving.

* Initialize dotenv in the deploy script and check it to determine if we want to clean local cache post-build.

* Check cwd specifically for heroku

* List the public images cache for heroku

* Make down/uploads synchronous again

* Limit page builds

* Fix re-syncifying downloads/uploads

* Add root-level 404.html to gitignore

* Fix prune-cache to work post-Models

It was firing falsely because some pages still have slashes.
This commit makes the path set that cache is run against remove all
trailing slashes, matching the outputof crawlPageData.

* Start logging

* Improve cache dir debug tools

* Consolidate deploy options into one env var.

The only deploy options env var is DEPLOY_OPTIONS, and it takes a
comma-separated string of build steps to run on the deploy.

A setting of all possible steps would be:
DEPLOY_OPTIONS="download,logTrees,logHashes,bailAfterLogs,build,upload,clean"

Having DEPLOY_OPTIONS unset gives the following default:
DEPLOY_OPTIONS="download,build,upload,clean"

If DEPLOY_OPTIONS is specified, it will only perform the provided steps. The
default options are disregarded.

logHashes makes a small console log of a hash describing the structure of the
directory trees of `.cache` and `public`. If any file is added, removed, or
renamed, this hash will change. File content is disregarded, however cache
busted files with content hashes in their filenames will effectively have
content-detection behavior.

* Change heroku hook to run on bash and save output

* Make deploy logging script log both to file and console

* Remove NM cache heroku test

* log pages

* log branch env var

* Revert sharp-related gatsby plugins to pre-Jobsv2 behavior to help with nonstandard caching

* Remove page logging from build script and re-enable prod cache warming.

* Re-async downloads and uploads

* Fix {} typo in downloadAll function

* Remove branch logging from onPostBuild

* Remove deploy-with-s3.sh and direct call the js variant again.

* Re-add node_modules cacheDirectories entry

I don't know if it's redundant with the buildpack, but I don't think it can hurt.

* Remove temporary deploy logs.

They can be re-added later if needed, possibly in another branch.

* Remove S3 testing utils import

oops!

* Test Heroku cache

* Don't specify cache directories

* Bump CircleCI cache version (?)

Co-authored-by: Pavel Grinchenko <psdcoder@gmail.com>
Co-authored-by: RMP <rogermparent@gmail.com>