Fix and improve cache pruning (#1209)
I had some testing behavior in the cache file that logged file removals without
actually doing them. This would be fine if I had marked it in a TODO or any
comment whatsoever.

Anyway, this commit fixes that.

I also changed the logic to use regular `fs` and `Promise.new` instead of
`fs.promises` to avoid the experimental warning. `upath` is used because
I believe that without it cache pruning won't work on Windows.