Let findChildWithSource accept external links
Some parts of the application still use the sidebar.json helpers to determine
their links, and when attempting to do so on external-type links broke the
build.

This change fixes that situation, letting the `findChildWithSource` function
adapt to input that isn't internal.

It's a bit ugly, but this whole branch is a bit ugly. The sidebar overhaul
should take care of this more cleanly in the future.