Change sidebar external icon link to match font color.
I made a new SVG based off the other one. It uses "currentColor" so it will take
on the color of its surrounding font by default, and also is fully transparent.

This new SVG is only used in the sidebar- other external link icons are still
the old hard-coded blue and white one.