Move Community data into GraphQL
This allows for better build-time processing, enabling things like
build-time querying for non-expired Events and Heroes.

Instead of importing the JSON file directly, Community content data is now
accessed through a custom React hook that wraps `useStaticQuery` and massages
the data into a more usable form.

Events and Heroes are given custom node types, and everything else in the
community json file is carried over in a unique "rest" node that exposes the
rest of the properties in JSON.

The Hero and Event components on the community page are adapted to the new
custom nodes, and the other Community components are successfully working
through the Rest node- they can be truly ported later or left as-is if they
don't need to change.