Integrate smooth scroll into regular link component
Instead of a separate component, the Link component now accepts a
`scrollOptions` object that passes options to the internal local fragment scroll
handler.

This means you can provide a `scrollOptions` object to a link to affect its
scrolling behavior, and this is compatible with links that could be of any type
as the `scrollOptions` object is only read when the `Link` wrapper detects a
local fragment.

This commit also removes special `Link` behavior/imports from the `Hero`
component.