Create some example content with the new custom card components
This is not intended to be final, but show off the general premise of
how using custom components for this kind of layout would look like.

The card and cards components are, on top of being fully stylable with CSS
modules, capable of using arbitrary Markdown images or elements as the card
"logo" when the `logo` prop is set to `true` (or anything truthy, like the
string `"true"`). This works around the complexity of working in custom image
capability by leaning on the one we already have in Markdown.

To make a card with a logo, set the "logo" prop to true and make sure the first
element in the card contains the image you want- this element will be set aside
in its own div, with the rest of the contents displayed beside it using `flex`

There are some quirks with Remark custom components, though:

- custom components aren't indented

- custom components must be contained to one line, including props.

- custom components can include Markdown, but there must be an empty line
  between the transition.

- self-closing tags aren't supported

Should we use custom components more often, we can migrate to MDX (I actually
already have a fully working branch with it that I made on my own time) which
handles this syntax much more comfortably because it's designed around the
concept. All our custom remark-based logic will cleanly port because MDX
actually uses Remark in its pipeline.