Properly detect when on docs home
This fixes an unreported issue (well, at least I haven't seen it) where the docs
sidebar wouldn't mark home as active when the user is on it.

This fix won't work if we decide to add children to the docs home later on, but
we currently don't, I'm pretty sure we have no plans to, and we can easily adapt
this fix to handle it if we ever do.