Give docs home a special bullet
This branch adds the fields "icon" and "style" to the allowed fields in
sidebar.json.

The "style" field applies a class from the sidebar module to the item.
Similarly, the "icon" field lets one define a custom icon on a menu item that
replaces the default `::before` bullet. Usable icons must be added as a
component (any source works, I used SVGR but an `img` can be used for raster
images. The class has to exist in the sidebar CSS module to do anything, but
using one that doesn't exist won't break the build.

Once an icon is defined, it can be used and shared with any amount of sidebar
entries anywhere by using its key in the const as the value of "icon". Items
with `icon` defined will have the default triangle bullet removed.

The home SVG changes color with the link because it's made with `currentColor`
and an actual element instead of the more limited `::before`.