Fix docs sidebar home highlight (#1312)
* Properly detect when on docs home

This fixes an unreported issue (well, at least I haven't seen it) where the docs
sidebar wouldn't mark home as active when the user is on it.

This fix won't work if we decide to add children to the docs home later on, but
we currently don't, I'm pretty sure we have no plans to, and we can easily adapt
this fix to handle it if we ever do.

* Fix TS issues

* Fix variable rename typo from ts fix

* Change path to be sourced from Gatsby

Beforehand, we were getting the value from the `location` global. This way, we
aren't subject to the whims of the browser and will get the canonical path every
time.