Special bullet/style for sidebar docs home (#1322)
* Properly detect when on docs home

This fixes an unreported issue (well, at least I haven't seen it) where the docs
sidebar wouldn't mark home as active when the user is on it.

This fix won't work if we decide to add children to the docs home later on, but
we currently don't, I'm pretty sure we have no plans to, and we can easily adapt
this fix to handle it if we ever do.

* Fix TS issues

* Fix variable rename typo from ts fix

* Change path to be sourced from Gatsby

Beforehand, we were getting the value from the `location` global. This way, we
aren't subject to the whims of the browser and will get the canonical path every
time.

* Give docs home a special bullet

This branch adds the fields "icon" and "style" to the allowed fields in
sidebar.json.

The "style" field applies a class from the sidebar module to the item.
Similarly, the "icon" field lets one define a custom icon on a menu item that
replaces the default `::before` bullet. Usable icons must be added as a
component (any source works, I used SVGR but an `img` can be used for raster
images. The class has to exist in the sidebar CSS module to do anything, but
using one that doesn't exist won't break the build.

Once an icon is defined, it can be used and shared with any amount of sidebar
entries anywhere by using its key in the const as the value of "icon". Items
with `icon` defined will have the default triangle bullet removed.

The home SVG changes color with the link because it's made with `currentColor`
and an actual element instead of the more limited `::before`.

* Fix TS

* Give leaf items circle bullets by default

* slight adjustment on leaf bullet positioning

* Fix regression where items don't highlight at build time

I sourced from the wrong place in getting the current path earlier.

* Fix TS

* Fix auto-merge issues

* Change styling

- Remove bullets on leaves
- Remove bold styling on home
- Change padding to make non-bullet items look a little better

* Switch rotation of default arrow bullets

This would probably be better done by changing the file, so I'll do that
afterward.

* Rotate default bullet image

Such that the default state doesn't depend on `animation`.

* Change styles

- Re-add leaf bullets, but smaller now
- Implement suggested movement changes to home icon
- Lock icon to text gray