Convert Glossary to YAML data file
YAML is much friendlier than JSON for embedded multiline string content, and
with a slight modification to the JSON file loader we can handle them the exact
same way and parse the YAML to JSON at build time to serve to the components.

With a regular data file instead of a JS module, HMR works for tooltips
now (mostly)!
We used to have to restart the dev server to update tooltips, and the previous
iteration of this branch also had caching issues. This solves that by binding
the Glossary data to a file, piggybacking on its `contentDigest`.

You still have to refresh the page to update tooltips though, but I think this
is more due to how they are implemented with state on the client side. Either
way, it's an improvement from before and eliminating the need to restart gets us
most of the possible time savings in this regard.

- Added YAML support to `json-files` model. I didn't rename it because I think
  `json-files` more accurately describes how files are parsed into plain JSON
  than something like `data-files`

- Updated Glossary model to use YAML source file

- Converted ES5 glossary to YAML

- Explicitly add `js-yaml` and run `yarn update`
  `js-yaml` is used within Gatsby, so it's already installed.
  However, we get two different versions without an `update`