Change Glossary to be sourced from Markdown files.
This lays the groundwork to transform them into a more elaborated-upon
construct called "Basic Concepts"- basically, the glossary is going to have
extra data for users who aren't satisfied with a tooltip.

For now, this just sets the data structure and fixes the original issue of
build-time tooltip contents. The rest will be handled in a couple other
individual issues.

This commit removes the explicit addition of `js-yaml` since we don't need to
use it in our code. It's still around as a dependency of other packages.