Build time tooltips (#1334)
* Add sourceNodes to the Gatsby APIs called by Models

* Change glossary to ES5

This lets it be used at build time without transpilation.

We could also use JSON, but this change is smaller and easier.

* Remove react-markdown and replace it with build-time HTML

This is similar to the recent Community refactor. A model parses the raw JS
glossary data, walks the tree parsing all Markdown to HTML, then exposes the
resulting processed glossary via GraphQL and a useStaticQuery hook.

The `Tooltip` components that used to use `react-markdown` now use
`dangerouslySetInnerHTML` to consume description data.

* Add Types for Glossary hook

* Explicitly add remark-rehype

* Fix glossary entry for DVC cache

Remark dislikes special elements broken across lines halfway, so this is needed
to get the link processsed.

* Convert Glossary to YAML data file

YAML is much friendlier than JSON for embedded multiline string content, and
with a slight modification to the JSON file loader we can handle them the exact
same way and parse the YAML to JSON at build time to serve to the components.

With a regular data file instead of a JS module, HMR works for tooltips
now (mostly)!
We used to have to restart the dev server to update tooltips, and the previous
iteration of this branch also had caching issues. This solves that by binding
the Glossary data to a file, piggybacking on its `contentDigest`.

You still have to refresh the page to update tooltips though, but I think this
is more due to how they are implemented with state on the client side. Either
way, it's an improvement from before and eliminating the need to restart gets us
most of the possible time savings in this regard.

- Added YAML support to `json-files` model. I didn't rename it because I think
  `json-files` more accurately describes how files are parsed into plain JSON
  than something like `data-files`

- Updated Glossary model to use YAML source file

- Converted ES5 glossary to YAML

- Explicitly add `js-yaml` and run `yarn update`
  `js-yaml` is used within Gatsby, so it's already installed.
  However, we get two different versions without an `update`

* Fix link line break in glossary yml

Remark doesn't like breaking specifically between `]` and `(` in a link, but
this should be fine.

* Fix Community model

* Change Glossary to be sourced from Markdown files.

This lays the groundwork to transform them into a more elaborated-upon
construct called "Basic Concepts"- basically, the glossary is going to have
extra data for users who aren't satisfied with a tooltip.

For now, this just sets the data structure and fixes the original issue of
build-time tooltip contents. The rest will be handled in a couple other
individual issues.

This commit removes the explicit addition of `js-yaml` since we don't need to
use it in our code. It's still around as a dependency of other packages.

* Remove dead code caught by linter

* Revert more JSON+YAML data file logic to JSON only

* Change from double to single quotes in basic concepts

* Remove remark-rehype

We no longer use it anywhere.

* Re-use childNodeCreator in the other Models.

This simple wrapper was introduced in the Community Model. I'm just using it in
the other models now where applicable.

`childNodeCreator` makes a function that acts like `createNode`, but adds in the
parent ID of the current node calls `createParentChildLink` between the two
nodes, and resolves when the prior two tasks do.

I invoke the creators in the "parent" models' APIs so the Models that consume
them get a nice function for creating child nodes. The consumer API calls that
use this helper don't even need to destructure `actions` anymore, as making a
child node is their only function most of the time.

* Replace single-quotes with backticks (code quotes)

* Empty commit to test automatic deploys