Actually use express correctly for empty responses, and add logs
It's my first time using express, and I didn't realize I had to call `send()`
and it was failing silently. On that note, we now have a console warning that
shows when the stars response is anything but 200, showing the code. This is
good enough for debugging without revealing too much about our proxy's
implementation.