Overhaul how GitHub stars are handled (#1363)
* Overhaul how stars are handled

- Uses `gatsby-source-github-api` to get GitHub data at build time

- Splits the GitHub Issues proxy API into one for Issues and Stars, sharing the
  query data. The new endpoint is `/api/github/stars`.

- Updates the GithubLine component to display static value at first, then update
  via the proxy as soon as possible to reflect any changes.

* Fix TS in stars hook

* Remove unused desctructure (unrelated)

I just happened to notice this was caught by ts-lint when I finished up with it.

* Make build-time github fetching gracefully fail with no token

The source plugin normally throws, so we exclude it without an API token.
Templates now source from an intermediary Gatsby node that is populated with
example data when no token is given.

Currently, with no token, this branch displays stars as 8888 and issues become
stuck at loading.

* Remove check for unchanged star count

* Actually use express correctly for empty responses, and add logs

It's my first time using express, and I didn't realize I had to call `send()`
and it was failing silently. On that note, we now have a console warning that
shows when the stars response is anything but 200, showing the code. This is
good enough for debugging without revealing too much about our proxy's
implementation.

* Move issues API endpoint to /github/issues

It used to be /github before /github/stars existed, but this makes more sense.

* Clean only .cache when first build attempt fails

This usually fixes the problem and also lets use re-use images.

Can be fixed in a more elegant manner later once I diagnose why this is
happening, but this should at least save us from catastrophic build times on
Heroku while I do so.

* Remove now-unused links from exclude-links.txt

* Make stars nullable and give the component a default placeholder

* Hide star count before attempting proxy update